pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "2d-to-3d-relief"
version = "0.1.0"
description = "Image-to-Flat 3D Relief + Layer Swap Planner"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "2d-to-3d-relief contributors" }]
dependencies = [
  "numpy>=1.26",
  "Pillow>=10.0",
  "typer>=0.12",
  "rich>=13.0",
  "pydantic>=2.7",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
gui = [
  "PySide6>=6.6",
  "matplotlib>=3.8",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "pre-commit>=3.7",
]

[project.scripts]
relief = "twod_to_threed_relief.cli:app"
relief-gui = "twod_to_threed_relief.ui.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
